requirements.txt
Flask==2.0.2
mysqlclient==2.1.0
Flask-MySQLdb==0.2.0
Flask-Cors==3.0.10
Werkzeug==2.2.2
gunicorn==20.1.0
python-dateutil==2.8.2


